Add SES and separate concerns
