+++ email/sender/baseOptions.ts
@@ -1,0 +1,15 @@
+const baseOptions = {
+  Source: "noreply@kitaplatz-zentrale.de",
+  Message: {
+    Body: {
+      Html: {
+        Charset: "UTF-8",
+      },
+    },
+    Subject: {
+      Charset: "UTF-8",
+    },
+  },
+};
+
+export default baseOptions;
